Avoid usage of truffle methods in classfile project
