+++ espresso/src/com.oracle.truffle.espresso.classfile/src/com/oracle/truffle/espresso/classfile/attributes/CodeAttribute.java
@@ -20,27 +20,16 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-
 package com.oracle.truffle.espresso.classfile.attributes;
 
-import static com.oracle.truffle.espresso.classfile.bytecode.Bytecodes.INVOKEDYNAMIC;
-import static com.oracle.truffle.espresso.classfile.bytecode.Bytecodes.JSR;
-import static com.oracle.truffle.espresso.classfile.bytecode.Bytecodes.JSR_W;
-import static com.oracle.truffle.espresso.classfile.bytecode.Bytecodes.MONITORENTER;
-import static com.oracle.truffle.espresso.classfile.bytecode.Bytecodes.MONITOREXIT;
-import static com.oracle.truffle.espresso.classfile.bytecode.Bytecodes.RET;
-
-import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.espresso.classfile.ClassfileParser;
 import com.oracle.truffle.espresso.classfile.ExceptionHandler;
-import com.oracle.truffle.espresso.classfile.bytecode.BytecodeStream;
 import com.oracle.truffle.espresso.classfile.descriptors.Name;
 import com.oracle.truffle.espresso.classfile.descriptors.ParserSymbols.ParserNames;
 import com.oracle.truffle.espresso.classfile.descriptors.Symbol;
 
 public final class CodeAttribute extends Attribute {
-
     public static final Symbol<Name> NAME = ParserNames.Code;
 
     private final int majorVersion;
@@ -56,12 +45,6 @@ public final class CodeAttribute extends Attribute {
 
     @CompilationFinal(dimensions = 1) //
     private final Attribute[] attributes;
-
-    private static final int FLAGS_READY = 0x1;
-    private static final int FLAGS_HAS_JSR = 0x2;
-    private static final int FLAGS_USES_MONITORS = 0x4;
-    private static final int FLAGS_HAS_INDY = 0x8;
-    @CompilationFinal byte flags;
 
     public CodeAttribute(Symbol<Name> name, int maxStack, int maxLocals, byte[] code, ExceptionHandler[] exceptionHandlerEntries, Attribute[] attributes, int majorVersion) {
         super(name, null);
@@ -105,50 +88,6 @@ public final class CodeAttribute extends Attribute {
             }
         }
         return null;
-    }
-
-    /**
-     * Returns true if this method uses the JSR/RET bytecodes.
-     */
-    public boolean hasJsr() {
-        return (getFlags() & FLAGS_HAS_JSR) != 0;
-    }
-
-    public boolean usesMonitors() {
-        return (getFlags() & FLAGS_USES_MONITORS) != 0;
-    }
-
-    public boolean usesIndy() {
-        return (getFlags() & FLAGS_HAS_INDY) != 0;
-    }
-
-    private byte getFlags() {
-        byte localFlags = flags;
-        if (localFlags == 0) {
-            CompilerDirectives.transferToInterpreterAndInvalidate();
-            flags = localFlags = computeFlags(originalCode);
-            assert localFlags != 0;
-        }
-        return localFlags;
-    }
-
-    private static byte computeFlags(byte[] code) {
-        BytecodeStream bs = new BytecodeStream(code);
-        int bci = 0;
-        int flags = FLAGS_READY;
-        while (bci < bs.endBCI()) {
-            int opcode = bs.opcode(bci);
-            switch (opcode) {
-                case JSR, JSR_W, RET ->
-                    flags |= FLAGS_HAS_JSR;
-                case MONITORENTER, MONITOREXIT ->
-                    flags |= FLAGS_USES_MONITORS;
-                case INVOKEDYNAMIC ->
-                    flags |= FLAGS_HAS_INDY;
-            }
-            bci = bs.nextBCI(bci);
-        }
-        return (byte) flags;
     }
 
     public LineNumberTableAttribute getLineNumberTableAttribute() {
